requirements.txt
Django==3.2.23
djangorestframework~=3.14.0
certifi~=2021.5.30
pip~=23.3.1
setuptools~=39.0.1